Linting rules update.
1. I remove the comment on the indent rule because it seems that it's
not longer throwing a lint error. The issue from the comment it's still
open but it seems that our issue is fixed.

2. I had added an exception for the max length rule. URLs that are
longer than 100 characters will not throw an error.

3. I have added a file with rules specifically for the React
application.

I tried to merge the ES6 rules to the `.eslintrc` file located in the
project root folder. When I did that, there were a bunch of errors
related to the strict mode. You don't need to add 'use strict'; at the
top of ES6 modules. When the ecmaScript `modules` property is set to
true, the linter complains that the `use strict'; statement s from the
`lib` folder should be removed. When `modules` is set to false, the
linter says that we need to add 'use strict'; to all JSX files.

So I decided to keep the rules that apply to both the engine and the
react app in the root folder. And I created a special file for ES6 rules
that will be applied only for the rules. Let me know if you don't like
this approach.

I also added the parser `babel-eslint` because it's the only one that
can parse code like this

```javascript
export default class SomeClass extends React.Component {
add = () => {
    ...
};
```